+++ database/migrations/2018_09_19_182356_create_fines_table.php
@@ -17,7 +17,7 @@ class CreateFinesTable extends Migration
             $table->increments('id');
             $table->integer('loan_id');
             $table->date('last_run_date');
-            $table->integer('fine_amount');
+            $table->decimal('fine_amount');
             $table->date('next_run_date');
             $table->timestamps();
         });
